+++ CHANGELOG.md
@@ -11,8 +11,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - [api] Allow the export of key data
 - [api] Allow building of client-repo packages
-- [core] Allow the export of key data
 - [client] Ported the Morio client to Go
+- [core] Allow the export of key data
 - [core] Allow preseeding key data
 - [core] Generate GPG key pair to sign packages
 - [core] Automatically build a client packages on initial setup
@@ -22,6 +22,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - [core] Allow building of client-repo packages
 - [core] Support preseeding of client templates
 - [core] Added support for the RESEED_ON_RELOAD feature flag
+- [core] Added support for Debian client packages for amd64 and arm64
 - [dbuilder] Maintain APT repository of client packages
 - [ui] Allow the export of key data
 - [ui] Allow loading key data file in setup wizard